requirements.txt
aiofiles==0.3.1
httptools==0.0.9
invoke==0.15.0
Jinja2==2.9.6
MarkupSafe==1.0
pkg-resources==0.0.0
pypugjs==4.2.0
PyYAML==3.12
sanic==0.5.0
sanic-jinja2==0.5.1
six==1.10.0
treelib==1.3.5
ujson==1.35
uvloop==0.8.0
websockets==3.3
